JWT TOKEN: proper authorization

+++ src/app/api/publishedcollection/[collectionId]/route.ts
@@ -1,3 +1,4 @@
+import { validateAccessToken } from "@/lib/ValidateAccessToken";
 import dbConnect from "@/lib/mongo/dbConnect";
 import Collection from "@/models/Collection";
 import PublishedCollection from "@/models/PublishedCollection";
@@ -15,7 +16,7 @@ export async function GET(_request: NextRequest, { params }: { params: { collect
             return Response.json(publishedCollection)
         }
         else {
-            return new Response(JSON.stringify("Коллекция не найдена"), {
+            return new Response(JSON.stringify("Collection not found"), {
                 headers: {
                     "Content-Type": "application/json",
                 },
@@ -28,22 +29,22 @@ export async function GET(_request: NextRequest, { params }: { params: { collect
             headers: {
                 "Content-Type": "application/json",
             },
-            status: 404,
+            status: 500,
         })
     }
 }
 
-export async function DELETE(_request: NextRequest, { params }: { params: { collectionId: string } }) {
+export async function DELETE(request: NextRequest, { params }: { params: { collectionId: string } }) {
     try {
+
+        const result = await validateAccessToken(request)
+        if (typeof result !== 'string') return result
+
         dbConnect()
         const publishedCollection: PublishedCollectionType | null = await PublishedCollection.findByIdAndDelete(params.collectionId)
-        if (publishedCollection) {
-            await Collection.findByIdAndUpdate(publishedCollection.originCollection, { $set: { publishedCollectionRef: null } })
-            await User.findByIdAndUpdate(publishedCollection.authorId, { $pull: { publishedCollections: publishedCollection._id } })
-            return Response.json(publishedCollection)
-        }
-        else {
-            return new Response(JSON.stringify("Коллекция не найдена"), {
+
+        if (!publishedCollection) {
+            return new Response(JSON.stringify("Collection not found"), {
                 headers: {
                     "Content-Type": "application/json",
                 },
@@ -51,12 +52,37 @@ export async function DELETE(_request: NextRequest, { params }: { params: { coll
             })
         }
 
+        await Collection.findByIdAndUpdate(publishedCollection.originCollection, { $set: { publishedCollectionRef: null } })
+
+        const user = await User.findById(publishedCollection.authorId)
+        if (!user || user.email !== result) {
+            return Response.json("User not found", {
+                headers: {
+                    "Content-Type": "application/json",
+                },
+                status: 400,
+            })
+        }
+
+        user.publishedCollections.pull(publishedCollection._id)
+        user.save()
+        // await User.findByIdAndUpdate(publishedCollection.authorId, { $pull: { publishedCollections: publishedCollection._id } })
+
+        return new Response(JSON.stringify("Collection successfully deleted"), {
+            headers: {
+                "Content-Type": "application/json",
+            },
+            status: 200,
+        })
+
+
+
     } catch (e) {
         return new Response(JSON.stringify(e), {
             headers: {
                 "Content-Type": "application/json",
             },
-            status: 400,
+            status: 500,
         })
     }
 }
